Crate zng_wgt_shortcut

Shortcut display widget.

Crate

This crate is part of the zng project.

Structs

ChordSeparatorFnArgs

Arguments for chord_separator_fn.

KeyFnArgs

Arguments for key_fn.

KeyGestureFnArgs

Arguments for key_gesture_fn.

KeyGestureSeparatorFnArgs

Arguments for key_gesture_separator_fn.

ModifierFnArgs

Arguments for modifier_fn.

NoneFnArgs

Arguments for none_fn.

PanelFnArgs

Arguments for panel_fn.

ShortcutFnArgs

Arguments for shortcut_fn.

ShortcutText

W Display keyboard shortcuts.

ShortcutsSeparatorFnArgs

Arguments for shortcuts_separator_fn.

Statics

CHORD_SEPARATOR_FN_VAR

Widget function that generates the separator between key gestures in chord shortcuts.

FIRST_N_VAR

Maximum number of shortcuts to display.

KEY_FN_VAR

Widget function that generates the key view.

KEY_GESTURE_FN_VAR

Widget function that generates the key gesture panel.

KEY_GESTURE_SEPARATOR_FN_VAR

Widget function that generates the separators between modifiers and keys in a key gesture.

MODIFIER_FN_VAR

Widget function that generates the modifier view.

NONE_FN_VAR

Widget function that generates content when there is no gesture to display.

PANEL_FN_VAR

Widget function that generates the outer panel.

SHORTCUTS_SEPARATOR_FN_VAR

Widget function that generates the separator between shortcuts.

SHORTCUT_FN_VAR

Widget function that generates a shortcut panel.

Functions

chord_separator_fn

P Widget function that converts ChordSeparatorFnArgs to a widget.

default_chord_separator_fn

Default value for CHORD_SEPARATOR_FN_VAR.

default_key_fn

Default value for KEY_FN_VAR.

default_key_gesture_separator_fn

Default value for KEY_GESTURE_SEPARATOR_FN_VAR.

default_modifier_fn

Default value for MODIFIER_FN_VAR.

default_panel_fn

Default value for PANEL_FN_VAR.

default_shortcuts_separator_fn

Default value for SHORTCUTS_SEPARATOR_FN_VAR.

first_n

P Maximum number of shortcuts to display.

key_fn

P Widget function that converts a KeyFnArgs to a widget.

key_gesture_fn

P Widget function that converts KeyGestureFnArgs to a widget.

key_gesture_separator_fn

P Widget function that converts KeyGestureSeparatorFnArgs to a widget.

key_txt

Gets the localized key name.

keycap

Widget used b the default_modifier_fn and default_key_fn to render a Text! styled to look like a keycap.

modifier_fn

P Widget function that converts a ModifierFnArgs to a widget.

modifier_txt

Gets the localized modifier name.

none_fn

P Widget function that converts NoneFnArgs to a widget.

panel_fn

P Widget function that converts PanelFnArgs to a widget.

shortcut

c Shortcut(s) that must be displayed

shortcut_fn

P Widget function that converts ShortcutFnArgs to a widget.

shortcuts_separator_fn

P Widget function that converts ShortcutsSeparatorFnArgs to a widget.